+++ src/decoder_managed/animation.rs
@@ -99,6 +99,8 @@ impl ManagedAvifDecoder {
         Ok(DecodedAnimation {
             frames,
             info: DecodedAnimationInfo {
+                exif: self.exif_for(super::metadata::MetadataSource::Animation)?,
+                xmp: self.xmp_for(super::metadata::MetadataSource::Animation)?,
                 frame_count,
                 loop_count: anim_info.loop_count,
                 hdr: anim_info.hdr,
@@ -218,6 +220,8 @@ impl AnimationDecoder {
         };
 
         let info = DecodedAnimationInfo {
+            exif: inner.exif_for(super::metadata::MetadataSource::Animation)?,
+            xmp: inner.xmp_for(super::metadata::MetadataSource::Animation)?,
             frame_count: anim_info.frame_count,
             loop_count: anim_info.loop_count,
             hdr: anim_info.hdr,